chore: split the init method
